Changes to split party implementation - Supports playing arbitrary sound along with the split (rather than just teleport sound or no sound) - Correctly imported from older save files - Support for splitting off an arbitrary subgroup of the party rather than just a single PC (though the special node doesn't yet allow this) - Support for leaving town while split - if you leave town, the absent PC's items are not dropped on the ground when later entering a town - Option to not change location when reuniting - If you reunite in a different town than you started, you are returned to the town you split from

+++ src/classes/universe.h
@@ -43,7 +43,7 @@ class cCurTown {
 public:
 	bool quickfire_present = false, belt_present = false;
 	// formerly current_town_type
-	short num; // 200 if outdoors (my addition)
+	size_t num; // 200 if outdoors (my addition)
 	short difficulty;
 	bool hostile;
 	cPopulation monst;
